refactor(desktop/preload): move composition into preload/index.ts
Move the ElectronAPI composition (interface + value + contextBridge.expose
side effect) from preload.ts into preload/index.ts. preload.ts becomes a
two-line shim that imports the index for its side effect and re-exports the
ElectronAPI type, kept around so the existing esbuild entry point and
BrowserWindow webPreferences.preload path stay untouched.

The split is now complete: preload/ contains 14 files (13 namespace modules +
context-bridge.ts + ipc-listener.ts + types.ts + index.ts), each a small
focused module sourcing channel literals from `@shiranami/contracts`. The
preload.ts shim weighs 9 lines.
